a lot of stuff....

+++ ExtraBees2/src/sql/createMacroNetworkflow.sql
@@ -7,9 +7,9 @@ CREATE TABLE macro_networkflow (
   ipLayerProtocol int(11) NOT NULL,
   SumTotalBytesSrc bigint(20) DEFAULT NULL,
   SumTotalBytesDest bigint(20) DEFAULT NULL,
-  SumPacketSrc int(11) DEFAULT NULL,
-  SumPacketDest int(11) DEFAULT NULL,
+  SumPacketSrc bigint(20) DEFAULT NULL,
+  SumPacketDest bigint(20) DEFAULT NULL,
   SumDuration bigint(20) DEFAULT NULL,
-  SumConnections int(11) DEFAULT NULL,
+  SumConnections bigint(20) DEFAULT NULL,
   PRIMARY KEY (Starttime,srcESite,destESite,ipLayerProtocol)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8$$
